+++ packages/gatsby-source-wordpress-experimental/src/steps/source-nodes/before-change-node/menu.js
@@ -1,52 +1,45 @@
-export const menuBeforeChangeNode = async ({
-  remoteNode,
-  actionType,
-  wpStore,
-  fetchGraphql,
-  helpers,
-  actions,
-  buildTypeName,
-}) => {
-  if (
-    actionType !== `UPDATE` &&
-    actionType !== `CREATE_ALL` &&
-    actionType !== `CREATE`
-  ) {
-    // no need to update child MenuItems if we're not updating an existing menu
-    // if we're creating a new menu it will be empty initially.
-    // so we run this function when updating nodes or when initially
-    // creating all nodes
-    return null
-  }
+import PQueue from "p-queue"
+
+const menuItemFetchQueue = new PQueue({
+  concurrency: Number(process.env.GATSBY_CONCURRENT_DOWNLOAD ?? 200),
+  carryoverConcurrencyCount: true,
+})
+
+const fetchChildMenuItems = api => async () => {
+  const {
+    remoteNode,
+    wpStore,
+    fetchGraphql,
+    helpers,
+    actions,
+    buildTypeName,
+    additionalNodeIds,
+  } = api
 
   if (
-    (!remoteNode.menuItems ||
-      !remoteNode.menuItems.nodes ||
-      !remoteNode.menuItems.nodes.length) &&
-    (!remoteNode.childItems ||
-      !remoteNode.childItems.nodes ||
-      !remoteNode.childItems.nodes.length)
+    !remoteNode?.menuItems?.nodes?.length &&
+    !remoteNode?.childItems?.nodes?.length
   ) {
     // if we don't have any child menu items to fetch, skip out
-    return null
+    return
   }
 
   const selectionSet = wpStore.getState().remoteSchema.nodeQueries.menuItems
     .selectionSet
 
-  const query = `
-        fragment MENU_ITEM_FIELDS on MenuItem {
-          ${selectionSet}
-        }
+  const query = /* GraphQL */ `
+    fragment MENU_ITEM_FIELDS on MenuItem {
+      ${selectionSet}
+    }
 
-        query {
-            ${(remoteNode.menuItems || remoteNode.childItems).nodes
-              .map(
-                ({ id }, index) =>
-                  `id__${index}: menuItem(id: "${id}") { ...MENU_ITEM_FIELDS }`
-              )
-              .join(` `)}
-          }`
+    query {
+      ${(remoteNode.menuItems || remoteNode.childItems).nodes
+        .map(
+          ({ id }, index) =>
+            `id__${index}: menuItem(id: "${id}") { ...MENU_ITEM_FIELDS }`
+        )
+        .join(` `)}
+    }`
 
   const { data } = await fetchGraphql({
     query,
@@ -54,28 +47,19 @@ export const menuBeforeChangeNode = async ({
 
   const remoteChildMenuItemNodes = Object.values(data)
 
-  const additionalNodeIds = remoteChildMenuItemNodes.map(({ id } = {}) => id)
+  remoteChildMenuItemNodes.forEach(
+    ({ id } = {}) => id && additionalNodeIds.push(id)
+  )
 
   await Promise.all(
     remoteChildMenuItemNodes.map(async remoteMenuItemNode => {
-      if (
-        remoteMenuItemNode.childItems &&
-        remoteMenuItemNode.childItems.nodes &&
-        remoteMenuItemNode.childItems.nodes.length
-      ) {
-        // recursively fetch child menu items
-        const { additionalNodeIds: childNodeIds } = await menuBeforeChangeNode({
+      // recursively fetch child menu items
+      menuItemFetchQueue.add(
+        fetchChildMenuItems({
+          ...api,
           remoteNode: remoteMenuItemNode,
-          actionType: `CREATE`,
-          wpStore,
-          fetchGraphql,
-          helpers,
-          actions,
-          buildTypeName,
         })
-
-        childNodeIds.forEach(id => additionalNodeIds.push(id))
-      }
+      )
 
       const type = buildTypeName(`MenuItem`)
 
@@ -91,6 +75,26 @@ export const menuBeforeChangeNode = async ({
       })
     })
   )
+}
+
+export const menuBeforeChangeNode = async api => {
+  if (
+    api.actionType !== `UPDATE` &&
+    api.actionType !== `CREATE_ALL` &&
+    api.actionType !== `CREATE`
+  ) {
+    // no need to update child MenuItems if we're not updating an existing menu
+    // if we're creating a new menu it will be empty initially.
+    // so we run this function when updating nodes or when initially
+    // creating all nodes
+    return null
+  }
+
+  let additionalNodeIds = []
+
+  menuItemFetchQueue.add(fetchChildMenuItems({ ...api, additionalNodeIds }))
+
+  await menuItemFetchQueue.onIdle()
 
   return { additionalNodeIds }
 }